requirements.txt
lmdb=1.3.0
h5py=2.10.0
imageio=2.16.0
matplotlib=3.3.4
mkl=2020.2
ninja=1.10.2
numpy=1.19.2
numpy-base=1.19.2
opencv-python=4.6.0.66
pandas=1.1.5
pillow=8.3.1
python=3.6.13
pytorch=1.7.1=py3.6_cuda11.0.221_cudnn8.0.5_0
scikit-build=0.16.3
scikit-image=0.17.2
scipy=1.5.2
setuptools=59.6.0
termcolor=1.1.0
thop=0.1.1
tk=8.6.12
torch-fidelity=0.3.0
torchvision=0.8.2
tqdm=4.63.0